backend/functions/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.8.0
Pillow>=10.4.0
numpy>=1.26.0
